Add custom alveare exceptions
All Marshmallow exceptions are covered along with 404.

+++ alveare/common/rest.py
@@ -1,6 +1,6 @@
 from flask import jsonify, make_response, request
 from flask.ext.login import current_user
-
+from alveare.common.exceptions import ClientError
 from alveare.common.database import DB
 from sqlalchemy import or_
 
@@ -12,20 +12,24 @@ def get_collection(model, serializer, query_filter=None):
     return jsonify(**{model.__pluralname__: serializer.dump(all_instances, many=True).data})
 
 def add_to_collection(model, deserializer, serializer):
-    new_instance = deserializer.load(request.form or request.json).data
+    request_data = request.form or request.json
+    new_instance = deserializer.load(request_data).data
     DB.session.add(new_instance)
     DB.session.commit()
-
     response = jsonify(**{model.__tablename__: serializer.dump(new_instance).data})
     response.status_code = 201
     return response
 
 def get_resource(model, instance_id, serializer):
-    instance = model.query.get_or_404(instance_id)
+    instance = model.query.get(instance_id)
+    if not instance:
+        raise ClientError(404)
     return jsonify(**{model.__tablename__: serializer.dump(instance).data})
 
 def update_resource(model, instance_id, update_deserializer, serializer):
-    instance = model.query.get_or_404(instance_id)
+    instance = model.query.get(instance_id)
+    if not instance:
+        raise ClientError(404)
     fields_to_update = update_deserializer.load(request.form or request.json).data
     for field, value in fields_to_update.items():
         if getattr(instance, field) != value:
@@ -36,6 +40,8 @@ def update_resource(model, instance_id, update_deserializer, serializer):
 
 def delete_resource(model, instance_id):
     instance = model.query.get_or_404(instance_id)
+    if not instance:
+        raise ClientError(404)
     DB.session.delete(instance)
     DB.session.commit()
